refactor: Split large data_types.rs test file into focused modules
- Split 676-line data_types.rs into 4 focused test modules:
  - numeric_types.rs: Integer, float, and decimal type tests (6 tests)
  - temporal_types.rs: TIME, TIMESTAMP, and INTERVAL type tests (6 tests)
  - spatial_types.rs: Geometric/spatial data type tests (7 tests)
  - text_types.rs: TEXT, VARCHAR, and character type tests (3 tests)

- Updated mod.rs to include the new test modules
- Removed the original monolithic data_types.rs file
- All tests continue to pass after refactoring

Closes #808
